refactor(analysis-types): extract maintainability DTO
